+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Not released
+
+### Features
+
+* Add `repl.block()`
+* Allow to customize `block` through call options
+* Allow to customize `from` through call options
+
 ## 0.1.1 (2024-07-30)
 
 ### Features